tools/etl/parcels-solr/requirements.txt
psycopg2-binary==2.9.10
pysolr==3.10.0
pyyaml==6.0.2